% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_flat_file.R
\name{read_flat_file}
\alias{read_flat_file}
\title{Read Flat File CSDR Submissions into R}
\usage{
read_flat_file(path)
}
\arguments{
\item{path}{Required. Path to the file file CSDR Submission (NOTE: The flat
file submission must be in xlsx format. Open the file in Excel and "Save as
type: Excel Workbook (*.xlsx)").}
}
\value{
Returns a flat_file list object. The first list object, \code{MetaData} is
the submission's metadata, the second list object, \code{Data} includes the
cost, hours, and unit data by WBS, To Date and At Completion, by Recurring
and Nonrecurring, and by Functional Element and Functional Category.
}
\description{
Read legacy Flat File CSDR Submissions into R.
}
